Write submit time to prevent posting too quickly
The Submitter would post immediately on boot-up regardless of when it
last posted. This commit writes the last "successful" (actually posted
or ratelimited) submit time to a file so that the Submitter can delay
the initial post accordingly.

The Fetcher and ratelimit.Flag timestamp file handling have been
refactored to use src.util.readline.