Fix Submitter trying private/non-existant users
If the Submitter chooses a private or non-existant user, all of their
postable links will be tried (and fail) needlessly. Instead, this commit
verifies that the user exists and is not private before choosing that
user to post.

This is basically a tradeoff between one preliminary instagram network
hit vs. potentially K failed hits if the user is private or
non-existant. It probably evens out on average.