Fix bot not linking to private accts if new cache
The lookup would not be performed if the first fetch was performed (ie,
if .should_fetch is True). This would result in the bot sometimes not
linking accounts that should have been linked. This commit resolves the
issue by adding a .private check after the first potential fetch_data
call.

I'm 93% sure this bug has existed since the instagram refactor.