Fix potential subreddit add debug logging
The Replier's was logging an incorrect number of to-add points for
potential subreddits (successfully summoned-to subreddits that are not
already crawled by the bot).

This also changes the threshold behavior so that potential subreddits
are added when the to-add count reaches the threshold (==) instead of
adding them when the count surpasses it (>).