Fix bot replying with empty metadata line
If the bot was ratelimited during a reply that did not require a fetch,
it would reply with an empty metadata line (ie: '[ ]'). This commit
fixes this by 1) ensuring that an empty metadata line is not committed
to the reply if that does somehow happen in the future and 2) sleeping
(with self._killed.wait) if a metadata fetch occurs while the bot is
instagram ratelimited. The second case was (in theory) already happening
if the metadata request itself encountered the ratelimit but not if the
bot was already ratelimited.

The rest of the changes are 1) fixing a 500-level request delay bug and
2) trying to harden the Fetcher against ratelimit handling bugs (namely
incorrectly NOT resetting the ratelimit variables when no longer
ratelimited which I still have not tracked down yet).