Fix bot sometimes linking pvt accts w/ 0 followers
The bot was sometimes linking to private accounts with zero followers
because the Fetcher.num_* properties were returning an 'A or B'
statement. Since zero is equivalent to False, the 'B' in the 'A or B'
statement was being returned. This ultimately resulted in the bot
thinking that the account had enough followers.

This issue has been fixed at both the Fetcher and Formatter levels.