Notify mention author if bot is banned from sub
Previously there was no notification given to the mention author if the
bot cannot reply. This commit gives the bot the ability to send a pm to
the mention's author if the bot is either banned from the subreddit or
has the subreddit blacklisted.

Integrating in more filtered cases such as not replying because it
already did; not replying because it has made too many posts in the
submission/thread; and not replying because the account was already
linked and is private are things I want to eventually add, but they
complicate the code considerably.

The reddit ratelimit handling for PMs has been hacked in: the title
column is repurposed for the PM's subject. This effectively removes the
mutual exclusion between the body column and title/selftext/url columns
from the RedditRateLimitQueueDatabase. Bugs related to this will
probably be a bit harder to debug because of this change. A better
implementation would probably involve separating the tables requiring
different columns (so: one table for replies, one table for submissions,
and one table for PMs).