Remove stale IG queue users
Stale users contained in the InstagramQueueDatabase were being updated
rather than removed. These users were essentially littering the database
since individual users are not re-visited often. Typically users that
get stuck in this database indicate a bug in the program.

Users are now deleted from the InstagramQueueData so that the database
is kept clean.