Fix? procs fetching same IG user simultaneously
There was a rare bug where multiple processes (ie, the Replier and
Submitter) could both attempt to fetch the same instagram user at the
same time. This never actually happened but this commit should prevent
this situation from ever occuring. Additionally it should also prevent
a secondary process from replying with expired data if a fetch of that
user is currently happening in another process.

The chance that this ever occurs is pretty low.