Implement gfycat album api handling
Gfycat's album api documentation is somewhat barebones. Thankfully,
their desktop site uses the api.

In order for an album to be fully functional, the
'/me/albums/{albumId}/name' needs to be hit to set the album link text
otherwise the album cannot be linked to and its contents will not behave
properly.

Created ._request_authed method to handled authorized requests. The
lower half of the diff is an indentation change.

I'm thinking to refactor the Uploader class to simply handle gfycat api
requests 1:1 and have the higher-level logic shifted out to the Gfycat
class. The Uploader class is currently doing too much, I think.