Fix replier Formatter crashing on private mismatch
The Formatter would crash if the instagram user database was flagged as
private while the live account was not.

This commit fixes the issue by fetching currently public account data
even if it is flagged as private, rendering the private flag completely
obsolete. Further, this introduces an extra (single) fetch for every
encountered instagram user (I think?).

I've also added a fallback (to the account being private) in the
Formatter to prevent it from crashing if it ever encounters a private
state mismatch in the future.