[INCOMPLETE] Implement PM highlights
This commit gives the bot the ability to reply with highlights to
private messages containing one or more instagram user account
names/links.

It has been tested with --dry-run and is working except for the fact
that the Replier does not properly re-fetch the enqueued Message. This
is because the reddit api does not have an endpoint for specific message
fetching.

The two solutions that I am entertaining are:
        1) expand the reply-queue database to store all of the relevant
                Message info & re-build the Message in the Replier from
                that data or
        2) fetch the message inbox stream (ideally from newest ->
                oldest) until the enqueued Message is found (or until
                the end of the stream in case the Message was deleted)
Currently, I am leaning toward 1) to avoid fetching the message inbox
stream but I am not sure if .reply will function properly with a
manually-built Message instance.