Fix non-existent users terminating fetcher process
If the _get_meta_data method fails because the user is non-existent, the
sanity check from ac07ea7 to prevent hitting /graphql/query without a
user-id was causing the process executing the fetch to terminate.

This has been fixed by checking if the user has been flagged as
non-existent before performing the user-id sanity check.