Allow bot to be resummoned to same post
The mentions database was disallowing the bot to be resummoned to the
same post because it was only storing the submission id. This was
causing the bot to miss some replies:
        - resummoned to post with new instagram data
        - resummoned to post after being unblacklisted/unbanned
                (extremely rare)

This commit fixes this issue by adding a new column to the mentions
database: mention_fullname. The database will now allow the bot to be
resummoned to duplicate posts.

Note: because there is no separate mentions-pm database, the pm logic
includes a workaround to prevent the bot from sending a duplicate pms to
mention authors (since all mentions will have to be re-parsed).