Add mentions-pm TODO
Added a TODO for PMing mention authors about status of summon. This
change requires a small(?) refactor in how the replies/Filter functions
which may or may not affect other parts of the bot.

Originally, the mentions system was written as an afterthought which was
a mistake since it is somewhat user-facing (but was rarely used when it
was written -- obviously).

I am extremely unmotivated to add new features to the bot right now. Bug
fixing is fine but refactoring does not excite me at all. Maybe that
will change later.