Fix April 6, 2018 graphql forbidden
Instagram seems to have invalidated the previous query_id. This
query_hash is a non-permanent solution as they will most likely
invalidate this as well (possibly soon) since it appears that these are
intended for instagram internal use.

See: https://stackoverflow.com/a/49270809

I'm not sure what to do when/if they change this (and if a new
query_{id/hash} does not leak).