Fix graphql query forbidden (try #2; ab28760)
My initial assumption that the query_id was invalidated was incorrect.
It seems that requests against the graphql/query endpoint now yield
forbidden if the 'ig_pr' cookie is not present. I have tested this in
the python interpreter so this fix _should_ work.

I'm not sure exactly where this cookie is set. It is buried somewhere in
one of their various minified javascript files. As to how often this
value changes, I have no clue.